+++ src/utils/chains.ts
@@ -38,6 +38,7 @@ export enum FEATURES {
   MULTI_CHAIN_SAFE_CREATION = 'MULTI_CHAIN_SAFE_CREATION',
   MULTI_CHAIN_SAFE_ADD_NETWORK = 'MULTI_CHAIN_SAFE_ADD_NETWORK',
   PROPOSERS = 'PROPOSERS',
+  TARGETED_SURVEY = 'TARGETED_SURVEY',
 }
 
 export const FeatureRoutes = {
